requirements.txt
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.2
django-cors-headers>=4.2.0



# Tools
python-dateutil>=2.8.2
Pillow>=10.0.0

# Production
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
factory-boy>=3.3.0